store beat reference as double

+++ matron/src/clock.h
@@ -10,8 +10,8 @@ typedef enum {
 void clock_init();
 bool clock_schedule_resume_sleep(int thread_id, double seconds);
 bool clock_schedule_resume_sync(int thread_id, double beats);
-void clock_update_reference(int beats, double beat_duration);
-void clock_update_reference_from(int beats, double beat_duration, clock_source_t source);
+void clock_update_reference(double beats, double beat_duration);
+void clock_update_reference_from(double beats, double beat_duration, clock_source_t source);
 void clock_set_source(clock_source_t source);
 void clock_cancel_all();
